corebackend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain==0.1.16
langchain-openai==0.1.7
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
duckduckgo-search==4.4.3
pydantic==1.10.13
gunicorn==21.2.0
python-jose==3.3.0
openai==1.30.5
httpx==0.27.0 